add azure-autorust bin
