+++ services/autorust/Cargo.toml
@@ -2,4 +2,5 @@
 members = [
   "codegen",
   "openapi",
+  "azure-autorust",
 ]
